- int not size_t for RI - Turn on KeyTrack on SinePlus (addresses #1808)

+++ src/scxt-core/tuning/midikey_retuner.h
@@ -45,7 +45,7 @@ struct MidikeyRetuner
 
     void setTuningMode(TuningMode);
 
-    size_t getRepetitionInterval() const;
+    int getRepetitionInterval() const;
 
     /*
      * Return the number of 12-tet smitones to retune by
